[IMP] l10n_ke_hr_payroll: add commuter and housing allowances and fixes
We add a commuter and housing allowance. The housing allowance is either a fixed
 amount or a percentage of the gross amount.

We also change the max food allowance rule parameter for December 2024.

This also makes some fixes, such as removing the employee's pin, which is a
duplicate of the KRA pin. Lastly, we reorder and rename some salary rules.

closes odoo/enterprise#89976

Task: 4922304
Related: odoo/upgrade#8278
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>