[IMP] mail,*: Make Store in mock server more generic
*: calendar, hr_holidays, im_livechat, rating, website,
website_livechat, website_slides, approvals, voip, whatsapp

The goal of this commit is to make the Store in the mock server more
like the actual store.

This allows to add generic models without _to_store methods to be added
to the store as well as allowing to follow more the same structure as
the actual server.

It adds support for:
- _to_store_defaults method on models
- _to_store now uses the current model
- add method was split and we can now use _add_records_field like in the
actual server
- Store.attr, Store.one, Store.many can now support lists, and other
Store.attr inherited object

task-4789853

closes odoo/enterprise#90046

Related: odoo/odoo#218376
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>