[FIX] website_helpdesk: prevent searchbar traceback on _onKeydown event
Steps to Reproduce:
1. Navigate to '/helpdesk/customer-care-1/knowledgebase'.
2. Enter any text in the search bar to trigger suggestions.
3. Press the down arrow key twice.
4. A traceback error occurs, indicating an Odoo Client Error.

Before this commit, we had a crash when a user types something in a
searchbar in the website helpdesk and do keydowns:
"Cannot read properties of undefined (reading 'nextElementSibling')"

To fix this, we need to add the 'data-bs-toggle' attribute to let BS
find it to set '_element' (which must be the previous sibling of the
dropdown-menu). This data attribute was probably not set before because
in this case the BS dropdown is not used on a button or a link as
documented; we use it as an autocomplete functionnality.

task-4752497

closes odoo/enterprise#97334

X-original-commit: 931779b845291e1dc7efdea91d91bdfa5ab703d6
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Darpan Tiku Daswani (dtda) <dtda@odoo.com>