[FIX] accountant: prevent displaying 0 cost items in XML reports
_______________________________________
On the export intrastat report page, we can click on the 'Posted
Entries' button, which allows us to select 'Hide lines at 0'. Doing
this will prevent 0 cost transactions from appearing in the
report. However, when exporting to XML, such transactions are still
displayed.
Note: this is related to Belgian companies only.
1. Go to settings and make sure the company you're operating with
is Belgian. Then, in accounting, under the Custmer Invoices
section, check the 'Intrastat' option.
2. Go to products and create a new product of type 'Goods'. Then,
in the 'Accounting' tab, click on a random commodity code. Set the
country of origin as a country belonging to the European Union
(for example, Austria). Click on save. You should see a field named
'Intrastat Supplementary Unit Amount' appear.
3. Go to accounting. Click on the 'Vendors' tab and 'Bills'. Click
on New. Select a random vendor, and add a line containing the
product you just created. Add the column 'Intrastat' and set the
corresponding field at '11 outright...'. Then, set a price.
In the 'Other Info' tab, set the field 'Intrastat Country' as a
random country in the European Union (for example, Australia). Set
the fiscal position at 'Import/Export' and set the 'Intrastat
Transport Mode' field at '1. Sea transport'. Finally, click on
confirm.
4. Duplicate the vendor bill you just created. On this new vendor
bill, set the Intrastat at '12 Direct trade...'. Display the column
'Disc.%' and set the field at 100. Then click confirm.
5. Click on the 'Reporting' tab, and under 'Audit Reports', click
on 'Intrastat Report'. Set the month to the current month, then
click on 'Report: Intrastat(Services....)' and select 'Intrastat
(Goods)'. You should be able to see the 2 vendor bills you just
created. Click on 'Options' then 'Hide lines at 0': you should see
the bill having a 0 cost disappear. Finally, click on the cog on
the top left next to 'Intrastat Report' and click on 'XML'.
6. Once the XML file has finished downloading, open it.

The transaction having a 0 cost shouldn't appear.

The transaction having a 0 cost appears.

In the code responsible for the display of the XML report, there
isn't any condition responsible for checking if the transaction
should be displayed or not.

_________________________________________
opw-4464123

---

closes odoo/enterprise#90537

X-original-commit: 0a66d02ce5d15394f44d6000843cc26dbb3981ac
Signed-off-by: Maeva de Keyser (mdek) <mdek@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>