[FIX] account_reports: mail template traceback
**PROBLEM**
In the accounting app, when selecting multiples customers, if one of them doesn't have a name, the cron sending the report will traceback.

**STEP TO REPRODUCE**
- On a clean db, install the account_reports module
- From the accounting app, create a new company customer (Customers/Customers)
- From that new company customer form view, add a new contact of type "other" and don't specify a name for it.
- Create an invoice for the company
- From the customer list view, select all, and trigger the action "Open Customer Statements"
- Send -> Print and Send
- Check the console, the cron task should traceback on the template

**CAUSE**
In accout_report.py, we get the name of each selected partners, and we don't check if the name doesn't exist.
In the template `pdf_export_filters`, we try to join all the name in a string, it fails because one of the name isn't a string (its value is False because it doesn't exist on the partner record).

**FIX**
When getting the partners names, filter out the partners without names.

opw-4916660

closes odoo/enterprise#91071

X-original-commit: 9cd7ac42162c78c242fb7f980f2b6a83f77ccf54
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Léo Gizard (legi) <legi@odoo.com>