[IMP] *: add voip.call form view.
*: voip, voip_{ crm, helpdesk, hr_recruitment, sale_subscription }

This commit adds the form view of the voip.call model in VoIP
+ several enhancements to the VoIP call management system. Key
changes include:

- Updated the `voip.call` model to inherit from both `mail.thread`,
allowing the form view to have a chatter.
- Added new computed/related fields such as `calls_count`,
`image_1920`, and `avatar_128` to improve the information available
for each call.
- Enhanced the `write` method to update the partner's phone number
when the `partner_id` is changed.
- Introduced a new `CallStatusBadgeField` component to visually
represent call statuses in the UI.
- Added buttons for viewing related calls, opportunities, applications,
tickets, and subscriptions directly from the call form.

Task-4909097

Part-of: odoo/enterprise#89781
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>