[IMP] stock_barcode: remove `markup` usage from barcode line description
This commit replaces the usage of unsafe `markup` with the CSS property:
`white-space: pre-wrap` for showing the line description in barcode.
Previously, `markup` was used to preserve the line breaks in the line
description by replacing all `\n`s with `<br/>` which required using
`markup` to render the text properly. However, the above mentioned CSS
property achieves the same exact result by not ignoring the line breaks
in the original string.

closes odoo/enterprise#90347

X-original-commit: 2c6f2a784913e9e7a1a27f20611c66bc37086e52
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Mohammad Abdulmoneim (abdu) <abdu@odoo.com>