[FIX] data_cleaning: prevent error when merging data
Currently an error occurs when we try to merge data in deduplication.

Steps to replicate:
- Go to: Data Cleaning > Configuration > Deduplication.
- Open any Deduplication Rules > Deduplicate > select any record
  (Make sure at least one record is available in list view).
- Click `Merge` on top left.

Error:
`ValueError: invalid literal for int() with base 10: 'NaN'`

The error because of a recent refactor where the data of a `many2one` will be an
object rather than an array. (task-3547961)

In `saas-18.3`, `record.data.group_id` (many2one field) was an array, so the
line [1] worked perfectly. But starting in `saas-18.4`, after the change, the
value became an object instead of an array. Because of that, `record.data.group_id[0]`
is now `undefined`, and `parseInt()` returns NaN, which causes the `int()`
on line [2] to throw a `ValueError`.

[1] - https://github.com/odoo/enterprise/blob/7d020d73762a17386cae4b68c15a5d59a51fb480/data_cleaning/static/src/views/data_merge_list_view.js#L108

[2] - https://github.com/odoo/enterprise/blob/7d020d73762a17386cae4b68c15a5d59a51fb480/data_cleaning/models/data_merge_group.py#L113

This commit solves the problem by accessing the ID using `.id` instead.

sentry-6739344848

closes odoo/enterprise#91190

X-original-commit: 57c05a7cb76e34161875bb18b34f773804a09ed3
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>
Signed-off-by: Bhavya Ashesh Nanavati (bhna) <bhna@odoo.com>