[FIX] pos_self_order_preparation_display: mobile
When doing a self order from a mobile device, the order was not sent to
the preparation display until it was paid at the cashier.

Steps to reproduce:
-------------------
* Setup a PoS with self-ordering mode set to 'mobile'.
* Setup a preparation display to show orders from this PoS.
* Place an order from a mobile device.
* Click on "Pay"
> Observation: The order does not appear on the preparation display.

Why the fix:
------------
Instead of just sending the order from the kiosk, we also need to
send it from the mobile device.

opw-4819732

closes odoo/enterprise#91394

X-original-commit: 22f8b35631070ca496aace71915e882d923d8987
Signed-off-by: Stéphane Vanmeerhaeghe (stva) <stva@odoo.com>
Signed-off-by: Robin Engels (roen) <roen@odoo.com>