[FIX] account_reports: Refresh checks when viewing them
Before, when opening a Return that had bypassed checks such as Draft
Entries, and we had deleted or confirmed those entries it wouldn't
have refreshed the check number and would still display that there are
draft entries.

Now, we are forcing the check to be refreshed when someone wants to
see them as to not crash the performances when someone opens the Tax
Returns Kanban view.

task-4850581

closes odoo/enterprise#93746

X-original-commit: bda85aa2db6fd20658de850ef13aee1d9fe87669
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Léo Leclerc (leol) <leol@odoo.com>