[FIX] account_accountant: with company cron
In this commit: https://github.com/odoo/enterprise/pull/89244/commits/788d79cea3c2d45582e16b25847c7821a6521da1
we fixed a bug where the cron would be trigger with the wrong company which
impacted the currency rate.
We decided to put the change in the try_auto_reconcile, but it was a bit of code
duplication so this commit will refactor that.

Task-4933083

closes odoo/enterprise#92459

X-original-commit: 22baeada443aee0bb13c49d5b2d63c1238d25f9a
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>