[IMP] point_of_sale: improve hoot system in PoS
Improved test hoots in PoS to test model and service functions. It is
now possible to access the `pos_store` via `getService(“pos”)`.

This will automatically load mocked data and information on the various
models.

Components can also be mounted independently of the rest of the
application, so that their functions can be tested.

For the moment, only `point_of_sale` models are loaded, but the
implementation is designed to be patched from other PoS modules.

---

A missing asset has been added to the html_editor module.

closes odoo/enterprise#90912

X-original-commit: e30e92f35d98e1c2ed4d5d204711afa86085b1b8
Related: odoo/odoo#220351
Signed-off-by: Stéphane Vanmeerhaeghe (stva) <stva@odoo.com>
Signed-off-by: David Monnom (moda) <moda@odoo.com>