[FIX] hr_payroll: empty journal entry
Problem
----------
When you create a draft entry from the payslip list view filtered by payrun, it creates a empty journal entry in Company US instead of the current one.

Objective
----------
Create a filled journal entry in the good company

Solution
----------
The context of the payrun card component was empty.
The context of the list controller is now transmitted to the Record Component to keep the same context

task-4932712

closes odoo/enterprise#92900

X-original-commit: 62ab43da387c2a2430cfe66babb34fa9f178eba8
Signed-off-by: Jurgen Gjini (jugj) <jugj@odoo.com>
Signed-off-by: Mathias Lebel (lebm) <lebm@odoo.com>