[FIX] account_disallowed_expenses_*: prevent duplicate line IDs
Before this commit:
When only the account_disallowed_expenses module was installed, the report
failed with a duplicate key error if a line had a 0 rate or no rate at all. This
was because such child lines were assigned the same line_id as their parent.

After this commit:
Lines with a 0 or no rate have the parent account appended to their line_id
to ensure uniqueness and avoid key collisions.

closes odoo/enterprise#91083

X-original-commit: 5801914e2a9b51070a3a2d929e45157156b660b5
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Harsh Trivedi (hatr) <hatr@odoo.com>