[IMP] documents_account: auto-configure journal sync folders
We enable the synchronization of document account by default, and we create
automatically the journal synchronization configuration along with the
corresponding folder and tag when a journal is created.

We also create all the synchronization configuration of the existing journal
when documents_account is installed.

Notes:
- we do the configuration of the journal synchronization in sudo because the
user creating the journal has not necessary access to document.
- we transfer the translation of the journal type to the folder name, but it
only transfers the installed language so to benefit from it, languages must be
activated before documents_account is installed.
- we also transfer the translation of the journal name to the documents tags.
But as we can only add translations after the creation of the journal where the
auto creation of the settings are performed, it will only benefit when
installing documents_account after having activated multiple language and added
translation to the journal names.

Task-4873102

Part-of: odoo/enterprise#90662
Related: odoo/odoo#219925
Related: odoo/upgrade#8105
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>