[FIX] hr_contract_salary: fix context on offer signature
When signing an offer, an employee and a version are generated. Some methods of the controller are using the active_test context, but without resetting to its default value.
This was causing some issue when creating or computing field of the versions. For example the current_version_id was computed to be an archive version, which is wrong.
This commit fixes the issue by adding the active field to the domain instead. The new version is also unarchived before archiving the old version to not trigger the version constraint about the active field.

Steps to reproduce:
- Create a job position and a candidate in recruitement.
- Generate an offer for the candidate.
- Sign and counter-sign.
- The current version on the newly created employee is wrong.

task-4926361

Part-of: odoo/enterprise#90412
Related: odoo/odoo#219400
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Jurgen Gjini (jugj) <jugj@odoo.com>