[FIX] pos_blackbox_be: restore `registerReceiptWeb` action
In this commit (f5cea73183c0a25039b44e8a16341095431de029), the `registerReceiptWeb` action was removed from the `pos_blackbox_be` module. This action was usefull to call `registerReceipt` and then call `send_blackbox_response`.
This commit restores the `registerReceiptWeb` action in the `BlackBoxDriver`, ensuring that when calling it, we send the blackbox response to the server (which is used for Kiosk flows).

closes odoo/enterprise#90293

X-original-commit: 1af28c6333bebc78203228c964f1ce2096774b1d
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>
Signed-off-by: Manu Vaillant (manv) <manv@odoo.com>