[IMP] sale_subscription: add prodata price toggle on services
Before this PR, Subscription Goods Products are always paid for the full
recurring period, even when added from an upsell in between
an invoicing period. For Subscription Service Products,
instead, we always calculate the prorated price when adding
them in between invoicing dates and we have the customer
only pay for the remaining days in the period. The behavior of the Goods
is intended and the flow is left untouched while for services, although
the default is still to calculate the prorated price, we leave the
option of deactivating this and invoicing for the full period price.

If one makes the prorata toggle visible on good type products as well
using studio, then whether the toggle is set or not, it doesn't matter,
it is ignored and the full amount is invoiced.

Also added a test.

closes odoo/enterprise#90268

Task: 4932997
Related: odoo/upgrade#8228
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>