[IMP] hr_payroll_account: set partner on payslip's move lines
Add the option on Salary rules to set the employee as the partner on the Journal Entry lines generated from the payslip. This works in both batched and non-batched move lines.

Adjust all the NET salary rules data to show partner per default.

This allows to track the chart of accounts through the employee's partner.

task-4652900

closes odoo/enterprise#81930

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>