[PERF] hr_referral: Send refusal notification to referrer's users
Description
-----------
Following odoo/enterprise@764f9b6f513c90eb79935d9e8469989ea704b61b,
`_send_notification` will not attach the message on the `hr.applicant`
itself, so a `model` was provided.
But due to a missing `res_id`, `message_notify` will set both
`model` & `res_id` to `False` since both are not specified. This
leads to the creation of a `mail.message` with no `res_id` and no
`model` (as `self` is an empty `mail.thread` recordset).

This leads to a performance regression when mail references need to
be computed to attach to the notification in
`_notify_by_email_get_base_mail_values`, due to the missing `model`
and `res_id` of the message.

This commit attaches the notification to the referrer's users instead,
providing its `model` and `res_id`.

Reference
---------
opw-4943488

X-original-commit: 3f419eabc28fcc6626b8b335e1b9422a0949379d
Part-of: odoo/enterprise#90684
Signed-off-by: Jurgen Gjini (jugj) <jugj@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>