[FIX] hr_referral: Do not send notif when opening the refuse wizard
Description
-----------
Currently, when opening the refusal wizard for an applicant, a call to
`archive_applicant` is made. However, there is an override of it in
`hr_referral` to send a notification for the applicant's refusal,
and closing the `referral_state`, even though the applicant may have
not been provided a reason for refusal, and the HR can simply close
the wizard instead of confirming it.

This commit removes the override of `archive_applicant` and moves
its logic into `action_refuse_reason_apply` which is called when
the wizard is confirmed.

Reference
---------
opw-4943488

closes odoo/enterprise#90684

X-original-commit: 58f3e3f7f4716b789ce7f1101c120fdd6c0ea089
Signed-off-by: Jurgen Gjini (jugj) <jugj@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>