[FIX] sign: fix sign item popup title after changing placeholder
Steps to reproduce:
- Create a document using a sign template.
- Add a new field or modify the placeholder of an existing one.
- Validate the changes.
- Notice title is replaced by the value entered in placeholder.

Issue:
After editing a template that has already been used in a signed document,
the title of each sign item popup is unexpectedly replaced with the value
 entered in the placeholder for that item. This corrupts the template and results
in incorrect title of sign item being shown when the template is reused.

Cause:
The issue occurs because the title was being updated with the placeholder
of that particular sign item.

Solution:
Added a prop for title of sign item in SignCustomPopover so that we can make the
title as name of the sign item or the actual type rather than the placeholder
of the sign item.

task-4872963

closes odoo/enterprise#91677

X-original-commit: eaea53e818b65eba08cffc6c8b91a9fe548f388e
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Devendra Jaikishan Ladhani (djal) <djal@odoo.com>