[IMP] sign{,_itsme}: show next documents to public signer
Before this commit, when signing the SENT requests through
a link (not logged in Odoo), the next documents to be signed
by that partner were not shown.

After this commit, the unlogged partner can now go to the
next documents after completing the signature of any SENT
request received.

task-4859714

closes odoo/enterprise#88639

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Co-authored-by: Arnaud Joset <arj@odoo.com>