[IMP] voip: show recent calls tab incase of missed calls
Purpose:
To open recent calls tab in softphone interface when there are missed calls.

Specification:
This commit updates the VoIP systray item behavior to ensure that,
when a user has missed calls, the softphone interface will automatically
switch to the "recent" calls tab.

Incase of an ignored incoming call, the softphone will now
switch to the "recent" tab instead of the "dialer".

Task-4917354

closes odoo/enterprise#90431

Signed-off-by: Louis Wicket (wil) <wil@odoo.com>