[IMP] mrp_mps: suggest and fill forecasted demand in mps
This commit enables the user to use a pop-up suggestion for the MPS.
The tool suggests for the user the forecasted demand, filling it out
in all periods for the chosen product in the MPS based on
historical or actual demand.

If the user chooses to apply the suggestion for a certain period and not
all the periods in MPS, the suggestion will be based only on the same
period but in the previous year.

Task-4827486

closes odoo/enterprise#90553

Related: odoo/odoo#222465
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>