[IMP] project_enterprise: plan tasks when creating project from template
This commit allows planning dates when a project is created from a template.
The dates of the created tasks are calculated based on the dates of the task in the template project.

Tasks will be scheduled based on the elapsed time between the start of the project and the start of the task in the template.
Tasks are scheduled using the smart scheduling algorithm, so the time ranges might not be exact to avoid conflicts, etc.

If the template doesn't have planned dates, tasks will be simply scheduled from the start of the created project (or today) using the smart scheduling.

Task-4700745

closes odoo/enterprise#91552

Related: odoo/odoo#221595
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>