[IMP] appointment: redesign calendar event form view
Enterprise counterpart to the redesign of the calendar
event form view.

Main changes (when appointment_type_id is set):

- Add ribbons for various appointment_status values
- Add buttons in the header to quickly update the status, with a
displayed according to the status, and associated actions on the model.
- Improve the 'resources' notebook page to display user or
resources depending on the type of appointment the meeting
is related to. This allows edition of booked capacity while
displaying relevant information.
- Remove unique child menu of Reporting, use the parent menu instead

COM PR odoo/odoo#219677
UPG PR odoo/upgrade#8160

Task-4929768

closes odoo/enterprise#90529

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>