[IMP] l10n_pe_edi: support prepayment in xml export
By default, the UBL module represents prepayments / downpayments as a
InvoiceLine in the final invoice with a negative quantity and amount.
However, SUNAT and the Peru Government does not allow negative
quantities, and instead require this information to be stored within the
PrepaidPayments node and referred to via AdditionalDocumentsReference.

task-4605204

closes odoo/enterprise#91396

X-original-commit: 019692aa88370b04f1a97af2f703c670c2ed1e7b
Related: odoo/odoo#221301
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Andrew Gavgavian (andg) <andg@odoo.com>