[FIX] social: fix incorrect getter for FieldPostPreview
This commit fixes an issue with a recent change done in
93aded56aa4f5623dfd4a984c4e52d1a9eac71a2. This commit changed the HtmlField used
from the old web_editor to the new html_editor.

Following this change the getter was wrongfully kept with the old name, which is
what this commit is correcting.

This commit also fixes an issue with link retargetting in social previews.
The issue is that the links inside the HtmlViewer are retargetted to follow a
specific structure which doesn't suit the social module needs.

It creates an extension of the HtmlViewer that replaces itself in the
components of FieldPostPreview. This extension replaces the retargetLink method
to an empty function.

task-4950837

closes odoo/enterprise#90391

Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>