[IMP] spreadsheet_edition: stop stringify pivot table
The values of the pivot table were always stringified with `toString()`.
This is not necessary anymore, now that we have registries to
normalize/sstringify the values.

The autofill expected the values to be stringified, so we need to
adapt it to the new behavior.

Task: 4266072
Part-of: odoo/enterprise#91679
Related: odoo/odoo#221845
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>