[IMP] documents_account: remove accounting sync option
The user can no longer disable the accounting synchronization with document
when documents_account is installed.

Note: Following the change of _get_default_folder in
account_reports.export.wizard where we just return the account_folder_id of the
company, we set precompute to True for the field account_folder_id of company
to ensure a folder is set by default. Note that the field is not required on
the model but well in the settings view. test failing without that change:
account_reports:TestAllReportsGeneration.test_generate_all_export_files

Task-4873102

Part-of: odoo/enterprise#90662
Related: odoo/odoo#219925
Related: odoo/upgrade#8105
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>