[FIX] mrp_mps: close edit production popup when opening BOM form
Before this commit:
-------------------------
Clicking the BOM field in the edit production schedule pop-up on the MPS page
redirected the user to the BOM form view, but the pop-up remained open in the
background. If the user then clicked Cancel on the pop-up, it created a
traceback error.

Steps to reproduce:
-------------------------
1. Install 'mrp_mps' module.
2. Go to Planning → Master Production Schedule.
3. Click the pencil icon next to a product.
4. In the pop-up, click the BOM value to navigate to the BOM.
5. Notice that the pop-up stays open even after redirection, and closing the
   pop-up raises a traceback error.

Cause of the issue
-------------------------
When your onClose arrow function directly executes
this.reload(productionScheduleId) without curly braces, it immediately hands
over the outcome of that function call to the dialog system. Since this.reload
is an asynchronous operation with multiple sequential return values, its
immediate outcome may not be a final value.

After this commit:
-----------------------
The pop-up is automatically closed when navigating to the BOM form view from the
BOM field in the Edit Production Schedule popup, ensuring a cleaner and more
intuitive user experience.

Task Id: 4886331

closes odoo/enterprise#94113

X-original-commit: d2749cb271fc603af6799fd926bbeba7ff03091a
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Harshil Patel (hapt) <hapt@odoo.com>