[FIX] helpdesk_stock: prevent traceback when returning products
step to reproduce :
create a sales order with a service and a consumable product
deliver the consumable product
enable the 'returns' feature on your helpdesk team
create a helpdesk ticket in this team
select the customer and the SOL you created
click on return
empty the delivery field => traceback

source of the issue :

Setting the delivery_id field to false triggers a compute, in which an
ensure one is later call on the delivery_id. Since the delivery is
empty, it triggers an error.

Solution :

The field 'delivery_id' is required. So we can simply prevent the
execution of the compute when the delivery is set to false and wait for
the user to put a new value to trigger the compute correctly.

task - 4935957

closes odoo/enterprise#91457

X-original-commit: d2c105ee55f170967d65fd622160341ce19e6b53
Related: odoo/odoo#221430
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>