[FIX] account_accountant: button on new partner
When creating a new partner from the set partner button, the partner doesn't
have any rank (supplier or customer), this lead to have the payable button to be
displayed since the condition to have it is to have a partner and not
(customer_rank > supplier_rank).
Since supplier rank and customer rank are 0 it will be False which lead to the
button to be displayed.
To know which button to display we check the amount of the statement line

no task id

closes odoo/enterprise#94852

X-original-commit: fd98234607c2f9191573bf57d90adc27950f2cf1
Signed-off-by: Ali Alfie (alal) <alal@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>