[FIX] website_appointment: fix online appointment option traceback
Since [1], the conversion of actions to classes introduced a forgotten
shared function, leading to a traceback when specifying appointment
types under the "Appointments" option.

Additionally, this commit corrects the missing `display_name` in
`SetAppTypesAction`. Previously, this omission caused the selected
appointment types to appear empty without a name in the UI.

Steps to reproduce:
- Drop "Appointment Button"
- Change "Appointments" option to "Specific Types"
- Click on "Specify" option and choose any record from the list
- Traceback occurs

[1]: 1f8b599a23330b03bffd898d85fe74a0f0eeda1f

closes odoo/enterprise#90717

X-original-commit: 8e9fffbe1cb87fb012fee6074e64df56920fe763
Signed-off-by: Serge Bayet (seba) <seba@odoo.com>