[FIX] stock_barcode_mrp: provide byProduct fallback value
Issue
-----
Trying to scrap a production through the barcode app gives the user a traceback.

Steps to reproduce
-----
- Have a done MO
- Go to barcode and try to scrap it

--> Traceback

Cause
-----
When preparing the context for the scrap window, we try to map the MO's
byProducts. When there is no byProduct, we end up trying to apply a mapping on
"nothing".

-----
Ticket:
opw-4792599

closes odoo/enterprise#93098

X-original-commit: 521bb12bb365437cd50648f8bf409010454a22e1
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: David Van Droogenbroeck (drod) <drod@odoo.com>