[IMP] documents{_account}: prevent document warning in the tests
In document._prepare_create_values, we emit a warning and set the owner to
False when the owner of the document to be created is not active. Now that we
have enabled the document account synchronization as soon as document_account
is installed, we get that warning in lot of account test.
To avoid those warning, we set upstream the owner to False when the owner is
not active.

Co-authored by: Stephane Debauche <std@odoo.com>

Task-4873102

Part-of: odoo/enterprise#90662
Related: odoo/odoo#219925
Related: odoo/upgrade#8105
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>