[ADD] pos_iot_adam_scale: move Adam driver to separate module
Before this commit, the Adam Equipment serial scale driver was included
by default on the IoT box. However, this causes issues as there is no
way to detect the presence of one of these scales automatically. This
caused any unidentified serial device to be registered as an Adam scale,
leading to much confusion with customers, especially regarding setting
up an FDM.

After this commit, the Adam driver is moved into its own module, and a
checkbox added to the PoS settings to enable it. This way, only
customers who actually need the driver will get it, and everyone else
can benefit from not having it wrongly detect devices.

task-4954021

closes odoo/enterprise#91410

Related: odoo/odoo#221326
Related: odoo/upgrade#8142
Signed-off-by: Max Whale (mawh) <mawh@odoo.com>