[FIX] l10n_se_sie4_import: Use UTF8 instead of ISO-8859-1
The aim of this commit is using UTF8 instead of ISO-8859-1
for the file decoding.
We received feedback from customers that the previous decoding
wasn't handle correctly swedish characters.
The documentation tends to say that we should follow CP437.
As UTF8 is a superset of CP437, we are now using it.

opw-4868415

closes odoo/enterprise#90870

X-original-commit: 32ab69c420d8a4949d0921334f618c8b88506874
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>