[FIX] ai: avoid duplicating the user prompt
Prior to this commit, we send messages to the llm like the following:
```
[
    {'content': Markup('<p>first question</p>'), 'role': 'user'},
    {'content': Markup("<p>Sure, I'm here to help. What's your first question?</p>"), 'role': 'assistant'},
    {'content': Markup('<p>second question</p>'), 'role': 'user'},
    {'role': 'system', 'content': "You are a RAG assistant.\n\nToday's date to be used: 2025-07-18"},
    {'role': 'user', 'content': 'second question'}
]
```

We're actually duplicating the user's prompt and this is because of the
retrieval of the chat history. Before calling generate_response, we post
the user's message. Therefore, his message is already recorded in the db.
We should then skip the most recent message to assemble the chat history
since its just the same to the prompt.

closes odoo/enterprise#90763

X-original-commit: 6c5c3bacd86a1ba65b91b9536bb6dbdd4ad55244
Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>