[FIX] helpdesk: prevent error when saving ticket without tags
Currently, an error occurs when the user tries to save the ticket after removing
the tag.

Steps to produce:
- Install 'Helpdesk' module.
- Create the 'Helpdesk team' with 'Automatic Assignment' as
  'Dispatch tickets based on tags'
- 'All Ticket' > Create a new ticket and set the Helpdesk Team as the newly
  created team and set any tag, and click save
- Remove the tag and save

Traceback:
IndexError: list index out of range

At [1], the error occurs because 'added_tags' is empty, which causes 'vals_list'
to be empty. As a result, attempting to unpack it using 'zip(*vals_list)' leads
to an IndexError.

[1]: https://github.com/odoo/enterprise/blob/6763c997bce8c5a42fd7c2b99d36ae8af5d8d238/helpdesk/models/helpdesk_ticket.py#L689

sentry-6754498068

closes odoo/enterprise#90727

X-original-commit: 3c2e61bb738663cdc5bd36567452f8cbc449642c
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Aayush Modi (aamo) <aamo@odoo.com>