[FIX] hr_contract_salary: fix salary configurator access issue
Since this commit: https://github.com/odoo/enterprise/commit/b68569868c00765bca910b5630a72db75347c03d, the salary configurator had an access issue on hr.job model.
This is due to the compute method on company_id introduced in the previous commit. As the properties definition field of the hr_job is defined on the company, it was recomputed every time the company commute is triggered.
Additionnaly, this compute on the company_id is useless as the domain on the contract_template_id is already filetered based on the company.

A test for this is added in this PR: https://github.com/odoo/enterprise/pull/90412

task-4951648

closes odoo/enterprise#90513

Signed-off-by: Jurgen Gjini (jugj) <jugj@odoo.com>