[FIX] knowledge: properly display comments with a top position = 0
Issue:
There is an issue in the logic of `computeVerticalDimensions` to display
comments. If the `top` value of a thread in the article is `0`, it will be
filtered out and not displayed because `top` was used as a boolean value.
Instead, it should properly consider `top` as a finite number to display the
comment or not.

Note:
This issue is not easily reproducible because there are few configuration where
a comment would have a top value equal to 0.

task-4984152

X-original-commit: d6962784b4d9205e10ab5c8ba2ac6847f11624f1
Part-of: odoo/enterprise#96047
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>