[FIX] knowledge: reload CommentBeaconManager when switching articles
Summary:
There was an issue where comments were not displayed when switching from a
locked article to another (read-only).

How to reproduce:
- Create two articles and add a comment on each.
- Lock both articles (so that they are effectively read-only).
- Switch from one article to the other.

Issue:
- Comments are not displayed to the user.

Resolution:
When switching between read-only articles, `KnowledgeHtmlViewer` is
not fully reloaded and continues using the same `CommentBeaconManager`
instance for the newly opened article. As a result, comment beacons
are not displayed when switching article. The simplest solution to this issue is
to re-instantiate a new `CommentBeaconManager` whenever the HTML value changes
to ensure comments are correctly displayed.

task-4984152

X-original-commit: f19896ae3697002b29666960982ec5d590f32741
Part-of: odoo/enterprise#96047
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>