[FIX] knowledge: preserve comment beacons located in baseContainer
Summary:
When a user adds a comment inside a baseContainer element, the comment
beacons created during the comment insertion can be discarded during
the document normalization step.

How to reproduce:
- Open an article in Knowledge.
- Select text and change the block style from "Paragraph" to "Normal"
  using the powerbox.
- Add a comment on the selected text using the powerbox.
- Write a message in the comment thread.
- Save and reload the article.

Issue:
- The comment beacons disappears from the editor and the user can't see it
  anymore.

Resolution:
When the editor is initialized, `div` are not yet categorized as
paragraph related elements. The `comments_plugin` logic to identify valid
positions for comments beacons should take that into account and allow elements
which are candidates to be a paragraph related element.

task-4984152

X-original-commit: 224879cc1d625bc5dd61c4ed71f8cf8694564495
Part-of: odoo/enterprise#96047
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>